+++ apps/commerce-api/src/main/kotlin/com/loopers/domain/product/ProductService.kt
@@ -1,9 +1,5 @@
-package com.loopers.domain.productlike
+package com.loopers.domain.product
 
-import com.loopers.domain.product.ProductCommand
-import com.loopers.domain.product.ProductEntity
-import com.loopers.domain.product.ProductRepository
-import com.loopers.domain.product.ProductSearchCondition
 import com.loopers.domain.vo.Price
 import com.loopers.support.error.CoreException
 import com.loopers.support.error.ErrorType
