+++ apps/commerce-api/src/main/kotlin/com/loopers/infrastructure/productlike/ProductLikeRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.product
+package com.loopers.infrastructure.productlike
 
 import com.loopers.domain.productlike.ProductLikeEntity
 import com.loopers.domain.productlike.ProductLikeRepository
