+++ apps/commerce-api/src/test/kotlin/com/loopers/domain/product/ProductServiceIntegrationTest.kt
@@ -3,7 +3,7 @@ package com.loopers.domain.product
 import com.loopers.domain.brand.BrandCommand
 import com.loopers.domain.brand.BrandEntity
 import com.loopers.domain.brand.BrandService
-import com.loopers.domain.productlike.ProductService
+import com.loopers.domain.product.ProductService
 import com.loopers.domain.vo.Price
 import com.loopers.support.error.CoreException
 import com.loopers.support.error.ErrorType
